LayoutTests/imported/w3c/web-platform-tests/navigation-api/per-entry-events/dispose-skip-current-on-truncate-expected.txt



FAIL Removing a currentEntry from the joint session history shouldn't dispose it assert_equals: expected 1 but got 2
